+++ acceptance-tests/src/testFixtures/java/tech/pegasys/teku/test/acceptance/dsl/TekuNode.java
@@ -19,6 +19,7 @@ import static java.util.stream.Collectors.toList;
 import static java.util.stream.Collectors.toSet;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.junit.jupiter.api.Assertions.assertEquals;
+import static org.junit.jupiter.api.Assertions.fail;
 import static tech.pegasys.teku.test.acceptance.dsl.metrics.MetricConditions.withLabelValueSubstring;
 import static tech.pegasys.teku.test.acceptance.dsl.metrics.MetricConditions.withNameEqualsTo;
 import static tech.pegasys.teku.test.acceptance.dsl.metrics.MetricConditions.withValueGreaterThan;
@@ -48,6 +49,7 @@ import java.util.Set;
 import java.util.concurrent.TimeoutException;
 import java.util.concurrent.atomic.AtomicReference;
 import java.util.function.Consumer;
+import java.util.function.Function;
 import java.util.function.Predicate;
 import java.util.stream.Collectors;
 import java.util.stream.IntStream;
@@ -95,6 +97,7 @@ import tech.pegasys.teku.spec.datastructures.state.beaconstate.BeaconState;
 import tech.pegasys.teku.spec.datastructures.state.beaconstate.versions.bellatrix.BeaconStateBellatrix;
 import tech.pegasys.teku.spec.logic.versions.capella.block.BlockProcessorCapella;
 import tech.pegasys.teku.spec.schemas.SchemaDefinitionsCapella;
+import tech.pegasys.teku.test.acceptance.dsl.Eth2EventHandler.PackedMessage;
 import tech.pegasys.teku.test.acceptance.dsl.GenesisGenerator.InitialStateData;
 import tech.pegasys.teku.test.acceptance.dsl.tools.BlsToExecutionChangeCreator;
 import tech.pegasys.teku.test.acceptance.dsl.tools.deposits.ValidatorKeystores;
@@ -152,8 +155,9 @@ public class TekuNode extends Node {
     container.start();
   }
 
-  public void startEventListener(final List<EventType> eventTypes) {
-    maybeEventStreamListener = Optional.of(new EventStreamListener(getEventUrl(eventTypes)));
+  public void startEventListener(final EventType... eventTypes) {
+    maybeEventStreamListener =
+        Optional.of(new EventStreamListener(getEventUrl(List.of(eventTypes))));
   }
 
   public void waitForContributionAndProofEvent() {
@@ -203,7 +207,7 @@ public class TekuNode extends Node {
 
   public void waitForBlockAtOrAfterSlot(final long slot) {
     if (maybeEventStreamListener.isEmpty()) {
-      startEventListener(List.of(EventType.head));
+      startEventListener(EventType.head);
     }
     waitFor(
         () ->
@@ -288,6 +292,52 @@ public class TekuNode extends Node {
     final String body = JsonUtil.serialize(signedBlsToExecutionChange, jsonTypeDefinition);
 
     httpClient.post(getRestApiUrl(), "/eth/v1/beacon/pool/bls_to_execution_changes", body);
+  }
+
+  public void waitForBlsToExecutionChangeEventForValidator(int validatorIndex) {
+    if (maybeEventStreamListener.isEmpty()) {
+      fail(
+          "Must start listening to events before waiting for them... Try calling TekuNode.startEventListener(..)!");
+    }
+
+    waitFor(
+        () -> {
+          final List<tech.pegasys.teku.api.schema.capella.SignedBlsToExecutionChange>
+              blsToExecutionChanges =
+                  getEventsOfTypeFromEventStream(
+                      EventType.bls_to_execution_change, this::mapBlsToExecutionChangeFromEvent);
+
+          final Optional<tech.pegasys.teku.api.schema.capella.SignedBlsToExecutionChange>
+              eventForValidator =
+                  blsToExecutionChanges.stream()
+                      .filter(m -> UInt64.valueOf(validatorIndex).equals(m.message.validatorIndex))
+                      .findFirst();
+          assertThat(eventForValidator).isPresent();
+        });
+  }
+
+  private <T> List<T> getEventsOfTypeFromEventStream(
+      final EventType type, final Function<PackedMessage, T> mapperFn) {
+    if (maybeEventStreamListener.isEmpty()) {
+      return List.of();
+    }
+
+    return maybeEventStreamListener.get().getMessages().stream()
+        .filter(packedMessage -> packedMessage.getEvent().equals(type.name()))
+        .map(mapperFn)
+        .collect(toList());
+  }
+
+  private tech.pegasys.teku.api.schema.capella.SignedBlsToExecutionChange
+      mapBlsToExecutionChangeFromEvent(final Eth2EventHandler.PackedMessage packedMessage) {
+    try {
+      return JSON_PROVIDER.jsonToObject(
+          packedMessage.getMessageEvent().getData(),
+          tech.pegasys.teku.api.schema.capella.SignedBlsToExecutionChange.class);
+    } catch (JsonProcessingException e) {
+      LOG.error("Failed to process bls_to_execution_change event", e);
+      return null;
+    }
   }
 
   private String getValidatorLivenessUrl(final UInt64 epoch) {