+++ networking/eth2/src/main/java/tech/pegasys/teku/networking/eth2/gossip/SignedContributionAndProofGossipManager.java
@@ -13,7 +13,6 @@
 
 package tech.pegasys.teku.networking.eth2.gossip;
 
-import java.util.Optional;
 import tech.pegasys.teku.infrastructure.async.AsyncRunner;
 import tech.pegasys.teku.networking.eth2.gossip.encoding.GossipEncoding;
 import tech.pegasys.teku.networking.eth2.gossip.topics.GossipTopicName;
@@ -45,7 +44,10 @@ public class SignedContributionAndProofGossipManager
         forkInfo,
         processor,
         schemaDefinitions.getSignedContributionAndProofSchema(),
-        Optional.empty(),
+        message ->
+            recentChainData
+                .getSpec()
+                .computeEpochAtSlot(message.getMessage().getContribution().getSlot()),
         maxMessageSize);
   }