+++ networking/eth2/src/test/java/tech/pegasys/teku/networking/eth2/gossip/AbstractGossipManagerTest.java
@@ -22,7 +22,6 @@ import static org.mockito.Mockito.verify;
 import static org.mockito.Mockito.verifyNoInteractions;
 import static tech.pegasys.teku.spec.config.Constants.GOSSIP_MAX_SIZE;
 
-import java.util.Optional;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import tech.pegasys.teku.infrastructure.async.AsyncRunner;
@@ -175,7 +174,7 @@ class AbstractGossipManagerTest {
           forkInfo,
           processor,
           gossipType,
-          Optional.empty(),
+          message -> UInt64.ZERO,
           maxMessageSize);
     }
   }