+++ CHANGELOG.md
@@ -23,5 +23,8 @@ For information on changes in released versions of Teku, see the [releases page]
 
 ### Additions and Improvements
 - Updated REST API to use Javalin 5
+- Tightened validation of the fork for gossip messages
+- Updated third party libraries.
+- Support for the `/eth/v1/beacon/blinded_blocks/{block_id}` REST API
 
 ### Bug Fixes